06-12-2023 2152

+++ MockitoTutorialForBeginners/src/test/java/com/in28minutes/powermock/PowerMockitoTestingPrivateMethodTest.java
@@ -25,7 +25,7 @@ public class PowerMockitoTestingPrivateMethodTest {
     public void powerMockito_CallingAPrivateMethod() throws Exception {
         when(dependencyMock.retrieveAllStats()).thenReturn(
                 Arrays.asList(1, 2, 3));
-        long value = (Long) Whitebox.invokeMethod(systemUnderTest,
+        long value = Whitebox.invokeMethod(systemUnderTest,
                 "privateMethodUnderTest");
         assertEquals(6, value);
     }
